Move Gate description into the project core and name the context test script
describeModule is a pure projection from a loaded module to its
description. It lived in the shell beside describeProject. Move it and
its types to project/core/description.ts, and keep only the loading
composition in the shell. Public names are unchanged.

The test:runtime script listed the four context test files under a
name from the old layout. Rename it test:contexts.
