+++ package.json
@@ -18,7 +18,7 @@
     "test": "node --disable-warning=ExperimentalWarning --experimental-strip-types --test tests/*.test.ts tests/composition/*.test.ts tests/composition/mutation/*.test.ts tests/adapters/*.test.ts tests/reporter/*.test.ts",
     "test:core": "node --disable-warning=ExperimentalWarning --experimental-strip-types --test tests/*.core.test.ts tests/self-hosted-policy.test.ts",
     "test:command": "node --disable-warning=ExperimentalWarning --experimental-strip-types --test tests/command.test.ts",
-    "test:runtime": "node --disable-warning=ExperimentalWarning --experimental-strip-types --test tests/project.test.ts tests/proof.test.ts tests/workspace.test.ts tests/run.test.ts",
+    "test:contexts": "node --disable-warning=ExperimentalWarning --experimental-strip-types --test tests/project.test.ts tests/proof.test.ts tests/workspace.test.ts tests/run.test.ts",
     "fixture:check": "node --disable-warning=ExperimentalWarning --experimental-strip-types packages/redproof/src/cli.ts check --config fixtures/eslint-project/redproof.config.ts",
     "fixture:prove": "node --disable-warning=ExperimentalWarning --experimental-strip-types packages/redproof/src/cli.ts prove --config fixtures/eslint-project/redproof.config.ts",
     "fixture:describe": "node --disable-warning=ExperimentalWarning --experimental-strip-types packages/redproof/src/cli.ts describe --config fixtures/eslint-project/redproof.config.ts",
