+++ packages/redproof/src/project/index.ts
@@ -1,4 +1,6 @@
 export type { LoadedGateModule, LoadedProject } from './core/discovery.ts';
+export type { DescribedProof, DescribedRule, GateDescription } from './core/description.ts';
+export { describeLoadedProject } from './core/description.ts';
 export { GateSelectionError } from './core/selection.ts';
 export { loadGateModule, loadProject, selectGateModules } from './shell/loader.ts';
-export * from './shell/describe.ts';
+export { describeProject } from './shell/describe.ts';
